# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StreamDestination {
  /** RTMP server URL (e.g. rtmps://live-api-s.facebook.com:443/rtmp/) */
  rtmpUrl: string;
  /** Stream key / token for the platform */
  streamKey: string;
}

export interface StreamConfig {
  /**
     * One or more RTMP destinations to stream to simultaneously
     * @minItems 1
     */
  destinations: StreamDestination[];
  /**
     * Optional source URL to relay (m3u8, YouTube Live, HLS, RTMP). If omitted, camera feed is used via WebSocket.
     * @nullable
     */
  sourceUrl?: string | null;
}

export type StreamSessionStatus = typeof StreamSessionStatus[keyof typeof StreamSessionStatus];


export const StreamSessionStatus = {
  idle: 'idle',
  connecting: 'connecting',
  live: 'live',
  error: 'error',
  stopped: 'stopped',
} as const;

export interface StreamSession {
  status: StreamSessionStatus;
  /** @nullable */
  rtmpUrl?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
}

export interface StreamError {
  error: string;
}