# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stream
    description: Live streaming operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stream/start:
    post:
      operationId: startStream
      tags: [stream]
      summary: Start a live stream
      description: Starts an FFmpeg process to push stream to RTMP endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamConfig"
      responses:
        "200":
          description: Stream started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamSession"
        "400":
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamError"
        "409":
          description: Stream already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamError"

  /stream/stop:
    post:
      operationId: stopStream
      tags: [stream]
      summary: Stop the current live stream
      responses:
        "200":
          description: Stream stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamSession"
        "404":
          description: No active stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamError"

  /stream/status:
    get:
      operationId: getStreamStatus
      tags: [stream]
      summary: Get current stream status
      responses:
        "200":
          description: Stream status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamSession"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    StreamDestination:
      type: object
      required:
        - rtmpUrl
        - streamKey
      properties:
        rtmpUrl:
          type: string
          description: RTMP server URL (e.g. rtmps://live-api-s.facebook.com:443/rtmp/)
        streamKey:
          type: string
          description: Stream key / token for the platform

    StreamConfig:
      type: object
      required:
        - destinations
      properties:
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/StreamDestination"
          minItems: 1
          description: One or more RTMP destinations to stream to simultaneously
        sourceUrl:
          type: ["string", "null"]
          description: Optional source URL to relay (m3u8, YouTube Live, HLS, RTMP). If omitted, camera feed is used via WebSocket.

    StreamSession:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [idle, connecting, live, error, stopped]
        rtmpUrl:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        error:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]

    StreamError:
      type: object
      required:
        - error
      properties:
        error:
          type: string