# lib/api-zod/src/generated/types/streamConfig.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StreamDestination } from './streamDestination';

export interface StreamConfig {
  /**
     * One or more RTMP destinations to stream to simultaneously
     * @minItems 1
     */
  destinations: StreamDestination[];
  /**
     * Optional source URL to relay (m3u8, YouTube Live, HLS, RTMP). If omitted, camera feed is used via WebSocket.
     * @nullable
     */
  sourceUrl?: string | null;
}