# lib/api-zod/src/generated/types/streamDestination.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StreamDestination {
  /** RTMP server URL (e.g. rtmps://live-api-s.facebook.com:443/rtmp/) */
  rtmpUrl: string;
  /** Stream key / token for the platform */
  streamKey: string;
}