# lib/api-zod/src/generated/types/streamSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StreamSessionStatus } from './streamSessionStatus';

export interface StreamSession {
  status: StreamSessionStatus;
  /** @nullable */
  rtmpUrl?: string | null;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
}